Fixed id=0 weight editing bug, and did a bit of refactoring
