+++ app/src/main/java/org/glucosio/android/activity/AddReadingActivity.java
@@ -1,0 +1,56 @@
+package org.glucosio.android.activity;
+
+import android.support.v7.app.AppCompatActivity;
+import android.widget.TextView;
+
+import com.wdullaer.materialdatetimepicker.date.DatePickerDialog;
+import com.wdullaer.materialdatetimepicker.time.RadialPickerLayout;
+import com.wdullaer.materialdatetimepicker.time.TimePickerDialog;
+
+import org.glucosio.android.R;
+import org.glucosio.android.presenter.AddReadingPresenter;
+import org.glucosio.android.tools.FormatDateTime;
+
+import java.text.DecimalFormat;
+import java.util.Calendar;
+
+public class AddReadingActivity extends AppCompatActivity implements TimePickerDialog.OnTimeSetListener, DatePickerDialog.OnDateSetListener {
+
+    private AddReadingPresenter presenter;
+
+    @Override
+    public void onTimeSet(RadialPickerLayout view, int hourOfDay, int minute, int seconds) {
+        TextView addTime = (TextView) findViewById(R.id.dialog_add_time);
+        DecimalFormat df = new DecimalFormat("00");
+
+        presenter.setReadingHour(df.format(hourOfDay));
+        presenter.setReadingMinute(df.format(minute));
+
+        Calendar cal = Calendar.getInstance();
+        cal.set(Calendar.HOUR_OF_DAY, hourOfDay);
+        cal.set(Calendar.MINUTE, minute);
+        FormatDateTime formatDateTime = new FormatDateTime(getApplicationContext());
+        addTime.setText(formatDateTime.getTime(cal));
+    }
+
+    @Override
+    public void onDateSet(DatePickerDialog view, int year, int monthOfYear, int dayOfMonth) {
+        TextView addDate = (TextView) findViewById(R.id.dialog_add_date);
+        DecimalFormat df = new DecimalFormat("00");
+
+        presenter.setReadingYear(year + "");
+        presenter.setReadingMonth(df.format(monthOfYear + 1));
+        presenter.setReadingDay(df.format(dayOfMonth));
+
+        String date = +dayOfMonth + "/" + presenter.getReadingMonth() + "/" + presenter.getReadingYear();
+        addDate.setText(date);
+    }
+
+    public void setPresenter (AddReadingPresenter newPresenter) {
+        this.presenter = newPresenter;
+    }
+
+    public AddReadingPresenter getPresenter() {
+        return this.presenter;
+    }
+}
